openapi/requirements.txt
# Networking
fastapi
uvicorn
# starlette==0.20.4

# Other
python-dotenv
loguru
pydantic
typing_extensions
python-keycloak
idna
llvmlite
numba
numpy==1.23.4
# mccabe==0.7.0
# click==8.1.3
# anyio==3.6.1
# idna==3.4
# h11==0.14.0
# flake8==5.0.4
# pycodestyle==2.9.1
# sniffio==1.3.0

# Monitoring
# sentry-sdk==1.9.*
# ddtrace==1.4.2

# DB
sqlalchemy==1.4.*
asyncpg==0.26.*
psycopg2-binary==2.9.*
alembic==1.*

# Tests
pytest==7.1.*
coverage==6.*
pytest-asyncio==0.19.*
pytest-httpx==0.21.*
